+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ Change Log
 - new ITER_STREAM_LINGER config item to keep streams alive for a given period after proxy disconnect (defaults to 30 sec.)
   @todo docs
 
+
 **Pyro 4.49**
 
 - added iterator item streaming support. It is now possible to return iterators from a remote